+++ masm-tasm/src/ASM/main.ts
@@ -1,5 +1,4 @@
 import * as vscode from "vscode";
-import { emulist } from "../emulators/main";
 import { nodejs_emu_list } from "../emulators/main-nodejs";
 import { ActionType } from "../utils/configuration";
 import { activateManager } from "./manager";
@@ -8,9 +7,7 @@ import * as statusBar from './statusBar';
 export async function activate(context: vscode.ExtensionContext) {
     statusBar.activate(context);
 
-    const execAction = activateManager(context, emulist.concat(
-        nodejs_emu_list,
-    ));
+    const execAction = activateManager(context, nodejs_emu_list);
 
     context.subscriptions.push(
         vscode.commands.registerCommand('masm-tasm.openEmulator', (uri: vscode.Uri) => execAction(ActionType.open, uri)),