+++ masm-tasm/src/emulators/main-nodejs.ts
@@ -1,7 +1,5 @@
-import { Dosbox } from "./dosbox";
-import { msdos } from "./msdos-player";
+import { JSDosHost } from "./jsdos";
 
 export const nodejs_emu_list = [
-    new Dosbox(),
-    new msdos(),
+    new JSDosHost
 ];